Pin manifest.
As discussed [1], this manifest should have been pinned i.e. the exact
commit hashes to be checked out are specified.

There is a little noise and reorganization in this commit over and
above simply adding the commit hashes.  This brings it inline with the
default output from the 'repo manifest -r' command.

[1] http://gumstix.8.x6.nabble.com/Simple-Build-of-Dizzy-for-Overo-td4969872.html

Signed-off-by: Ash Charles <ashcharles@gmail.com>