Freeze meta-ros temperately.
Building image is failed because of this repository and the error is "do_configure". Nothing we can do about it unless we spend time on this repository, which is a waste of time. In order to make build successfully, we freeze this repository. Keep an eye on it for further commits which may fix the error

Signed-off-by: Jerry Hung <jerry@gumstix.com>